# openapi/kurgin_backend_auth_openapi_v1.yaml
openapi: 3.1.0
info:
  title: KURGIN Backend Auth API
  version: 1.0.0
  description: |
    Backend contract for KURGIN auth, access, payments, Analyzer reports,
    Workspace collections, and Admin audit logs.
servers:
  - url: https://api.cvdrap.ru
    description: Production API placeholder
  - url: http://localhost:8000
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: Register user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Email already exists
  /auth/login:
    post:
      summary: Login user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          headers:
            Set-Cookie:
              description: Optional HttpOnly refresh token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
...
  /auth/me:
    get:
      summary: Current authenticated user
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Not authenticated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 10
        requested_role:
          type: string
          enum: [buyer, jeweler, designer, gemologist, partner]
          default: buyer
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          const: bearer
        user:
          $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
        plan:
          type: string
        access:
          type: object
          additionalProperties: true
...